refactor: aws 의존 파일 제거, 폴더 이름 수정
- migration 작업 중 aws 의존 문제 발생
- aws 의존 파일 제거
- 폴더 이름 수정
