--- backend/src/main/resources/application-docker-dev.yml
@@ -1,41 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://db:3306/dibb?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC&characterEncoding=UTF-8
-    username: root
-    password: dibb2024
-
-  jpa:
-    database-platform: org.hibernate.dialect.MySQLDialect
-    hibernate:
-      ddl-auto: none
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-        dialect: org.hibernate.dialect.MySQLDialect
-    defer-datasource-initialization: true
-
-  sql:
-    init:
-      mode: always
-      encoding: UTF-8
-
-server:
-  port: 8080
-  servlet:
-    encoding:
-      charset: UTF-8
-      force: true
-
-logging:
-  level:
-    org.hibernate.SQL: debug
-    org.hibernate.type: trace
-    org.springframework.security: debug
-
-security:
-  jwt:
-    header: Authorization
-    secret: c2Vqb25nLWRpYmItc3ByaW5nLWJvb3Qtand0LXNlY3JldC1rZXktc2Vqb25nLWRpYmItc3ByaW5nLWJvb3Qtand0LXNlY3JldC1rZXkK
-    token-validity-in-seconds: 86400
