--- docker-compose.yml
@@ -1,50 +1,0 @@
-services: #이 항목 밑에 실행하려는 컨테이너들을 정의
-
-  server:
-    container_name: back
-    image: ${DOCKER_USERNAME}/back:latest
-    expose:
-      - 8080
-    ports:
-      - 8080:8080
-
-  front:
-    container_name: front
-    image: ${DOCKER_USERNAME}/front:latest
-    expose:
-      - 3000
-
-  nginx:
-    container_name: nginx
-    image: nginx:latest
-    restart: unless-stopped
-    volumes:
-      - ./conf/nginx.conf:/etc/nginx/nginx.conf
-      - ./data/certbot/conf:/etc/letsencrypt
-      - ./data/certbot/www:/var/www/certbot
-    ports:
-      - 80:80
-      - 443:443
-    depends_on:
-      - server
-      - front
-#
-#  certbot:
-#    image: certbot/certbot
-#    restart: unless-stopped
-#    volumes:
-#      - ./data/certbot/conf:/etc/letsencrypt
-#      - ./data/certbot/www:/var/www/certbot
-#    entrypoint: "/bin/sh -c 'trap exit TERM; while :; do certbot renew; sleep 12h & wait $${!}; done;'"
-
-  mysql: # 서비스 명
-    image: mysql/mysql-server:8.0.23 #사용할 이미지
-    container_name: capo # 컨테이너 이름 설정
-    ports:
-      - "3306:3306" # 접근 포트 설정 ( 컨테이너 외부:컨테이너 내부)
-    restart: always
-    environment: # -e 옵션
-      MYSQL_DATABASE: # 데이터 베이스 이름
-      MYSQL_ROOT_HOST:  # MYSQL HOST
-      MYSQL_ROOT_PASSWORD:  # MYSQL 패스워드 설정 옵션
-      TZ: "Asia/Seoul"
